+++ AlgorithmInScala1.scala
@@ -52,7 +52,7 @@ object AlgorithmInScala1 {
    * iが0以下の場合、Noneを返却する。
    * 再帰版
    */
-  def fibonacci(i: Int): Option[Int] = {
+  def fibonacci(i: Int): Option[Long] = {
     // TODO: ロジックを実装
     None
   }
@@ -60,7 +60,7 @@ object AlgorithmInScala1 {
   /*
    * 末尾再帰版
    */
-  def fibonacci2(i: Int): Option[Int] = {
+  def fibonacci2(i: Int): Option[Long] = {
     // TODO: ロジックを実装
     None
   }
@@ -68,7 +68,7 @@ object AlgorithmInScala1 {
   /*
    * ループ版
    */
-  def fibonacci3(i: Int): Option[Int] = {
+  def fibonacci3(i: Int): Option[Long] = {
     // TODO: ロジックを実装
     None
   }
@@ -76,7 +76,7 @@ object AlgorithmInScala1 {
   /*
    * 再帰(メモ化)版
    */
-  def fibonacci4(i: Int): Option[Int] = {
+  def fibonacci4(i: Int): Option[Long] = {
     // TODO: ロジックを実装
     None
   }
@@ -86,7 +86,7 @@ object AlgorithmInScala1 {
    * nが負の数の場合、Noneを返却する。
    * 再帰版
    */
-  def factorial(n: Int): Option[Int] = {
+  def factorial(n: Int): Option[Long] = {
     // TODO: ロジックを実装
     None
   }
@@ -94,7 +94,7 @@ object AlgorithmInScala1 {
   /*
    * 末尾再帰版
    */
-  def factorial2(n: Int): Option[Int] = {
+  def factorial2(n: Int): Option[Long] = {
     // TODO: ロジックを実装
     None
   }
@@ -102,7 +102,7 @@ object AlgorithmInScala1 {
   /*
    * ループ版
    */
-  def factorial3(n: Int): Option[Int] = {
+  def factorial3(n: Int): Option[Long] = {
     // TODO: ロジックを実装
     None
   }
@@ -110,7 +110,7 @@ object AlgorithmInScala1 {
   /*
    * 畳み込み版
    */
-  def factorial4(n: Int): Option[Int] = {
+  def factorial4(n: Int): Option[Long] = {
     // TODO: ロジックを実装
     None
   }
@@ -118,7 +118,7 @@ object AlgorithmInScala1 {
   /*
    * 畳み込み版2
    */
-  def factorial5(n: Int): Option[Int] = {
+  def factorial5(n: Int): Option[Long] = {
     // TODO: ロジックを実装
     None
   }
